+++ connectx/__init__.py
@@ -1,3 +1,3 @@
-from .game import Game
-from .connectXEnv import ConnectXEnv
-from .learn import Learn
+from connectx.game.game import Game
+from connectx.env.connectXEnv import ConnectXEnv
+from connectx.agents.learn import Learn
