+++ connectx/agents/learn.py
@@ -1,6 +1,6 @@
 import os
 
-from connectx.connectXEnv import ConnectXEnv
+from connectx.env.connectXEnv import ConnectXEnv
 from stable_baselines3 import PPO, A2C
